demo: clean code with java enum

+++ spring-test/demo/src/main/java/com/gngsn/demo/cleanCodeWithEnum/Account.java
@@ -5,13 +5,13 @@ import java.math.BigDecimal;
 public class Account {
 
     private String user;
-    private double balance;
+    private long balance;
 
-    public void withdraw(double cash) {
+    public void withdraw(long cash) {
         this.balance = this.balance - cash;
     }
 
-    public void deposit(double cash) {
+    public void deposit(long cash) {
         this.balance = this.balance + cash;
     }
 }
